Refactor imports to replace 'libs.bootstrap.section' with 'libs.domain.section' for consistency across integration and command modules
