+++ integrations/discord/config.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Optional
 
 from integrations.base.interface import IntegrationsConfig
 from integrations.discord.events import comparison
-from libs.bootstrap.section import BaseSection
+from libs.domain.section import BaseSection
 
 if TYPE_CHECKING:
     from pathlib import Path  # noqa: F401
