+++ integrations/slack/config.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Any
 
 from integrations.base.interface import IntegrationsConfig
 from integrations.slack.events import comparison, slash
-from libs.bootstrap.section import BaseSection
+from libs.domain.section import BaseSection
 
 if TYPE_CHECKING:
     from pathlib import Path  # noqa: F401
