+++ libs/bootstrap/__init__.py
@@ -3,6 +3,5 @@
 
 - :doc:`libs.bootstrap.configuration`: 初期設定関数
 - :doc:`libs.bootstrap.app_config`: アプリケーション設定関連
-- :doc:`libs.bootstrap.section`: セクション情報取り込み
 - :doc:`libs.bootstrap.initialization`: 各種初期化処理
 """
