+++ libs/commands/registry/member.py
@@ -6,8 +6,8 @@ import logging
 from typing import TYPE_CHECKING, TypedDict, cast
 
 import libs.global_value as g
-from libs.bootstrap.section import BaseSection
 from libs.domain import modify, validator
+from libs.domain.section import BaseSection
 from libs.types import CommandType
 from libs.utils import dbutil, textutil
