+++ libs/commands/registry/team.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING, TypedDict, cast
 
 import libs.global_value as g
 from libs.bootstrap import initialization
-from libs.bootstrap.section import BaseSection
 from libs.domain import modify, validator
+from libs.domain.section import BaseSection
 from libs.types import CommandType
 from libs.utils import dbutil, formatter, textutil
