+++ libs/commands/report/entry.py
@@ -5,8 +5,8 @@ libs/commands/report/entry.py
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from libs.bootstrap.section import SubCommands
 from libs.commands.report import matrix, monthly, stats_list, stats_report, winner
+from libs.domain.section import SubCommands
 from libs.types import CommandType
 from libs.utils import dictutil
