+++ libs/commands/results/entry.py
@@ -5,8 +5,8 @@ libs/commands/results/entry.py
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from libs.bootstrap.section import SubCommands
 from libs.commands.results import detail, summary, versus
+from libs.domain.section import SubCommands
 from libs.types import CommandType
 from libs.utils import dictutil
