+++ libs/domain/section.py
@@ -1,5 +1,5 @@
 """
-libs/bootstrap/section.py
+libs/domain/section.py
 """
 
 import logging
